Extracted console extension
