+++ tests/Integration/ConfiguredReportsTest.php
@@ -3,7 +3,7 @@
 namespace PhpBench\Tests\Integration;
 
 use Generator;
-use PhpBench\Extension\CoreExtension;
+use PhpBench\Extension\ConsoleExtension;
 use PhpBench\Extension\ReportExtension;
 use PhpBench\Report\ReportManager;
 use PhpBench\Tests\IntegrationTestCase;
@@ -23,7 +23,7 @@ class ConfiguredReportsTest extends IntegrationTestCase
     public function testReport(string $generator, string $renderer): void
     {
         $manager = $this->container([
-            CoreExtension::PARAM_CONSOLE_OUTPUT_STREAM => $this->workspace()->path('test')
+            ConsoleExtension::PARAM_CONSOLE_OUTPUT_STREAM => $this->workspace()->path('test')
         ])->get(ReportManager::class);
         assert($manager instanceof ReportManager);
         $manager->renderReports(TestUtil::createCollection([
